requirements.txt
rdflib==4.2.2
PyYAML==5.1.1
html5lib